[parallel-2] done algorithm for slides 3-10
